Optmise Unit Testing
- Remove slow tests to its own excluded file so main unit tests run quickly
- Remove unnecessary removal of temp tables due to earlier Gravity Form conflict
- Run composer in production for automated tests (don't need dev tools as Travis comes with them)
- Rename groups / files to better match naming convention

- Move phpunit to the same folder structure as WP core
- Remove coveralls from automated testing as it takes too long
- Reorder automated tests so the bulk of the tests is done first (the quicker of the bunch)
