--- .coveralls.yml
@@ -1,3 +1,0 @@
-src_dir: .
-coverage_clover: ./tmp/clover.xml
-json_path: ./tmp/coveralls-upload.json
