+++ .travis.yml
@@ -25,12 +25,10 @@ matrix:
     env: WP_VERSION=3.9 WP_MULTISITE=0
 
 before_script:
-  - composer install --prefer-source --optimize-autoloader
+  - composer install --prefer-source --no-dev --optimize-autoloader
   - bash tests/bin/install.sh gravitypdf_test root '' localhost $WP_VERSION
 
 script:
- - phpunit -c phpunit.xml.dist --group ajax
- - phpunit -c phpunit.xml.dist
+ - phpunit -c tests/phpunit/phpunit.xml.dist
+ - phpunit -c tests/phpunit/phpunit.xml.dist --group ajax
 
-after_script:
-  - php vendor/bin/coveralls --verbose --exclude-no-stmt
