+++ tests/phpunit/phpunit.xml.dist
@@ -1,0 +1,38 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<phpunit
+	bootstrap="./bootstrap.php"
+	backupGlobals="false"
+	colors="true"
+	convertErrorsToExceptions="true"
+	convertNoticesToExceptions="true"
+	convertWarningsToExceptions="true"
+	verbose="true"
+	>
+	<testsuites>
+		<testsuite name="Gravity PDF Test Suite">
+			<directory suffix=".php">./unit-tests</directory>
+		</testsuite>
+	</testsuites>
+
+    <groups>
+        <exclude>
+            <group>ajax</group>
+        </exclude>
+    </groups>
+
+	<filter>
+		<whitelist>
+			<directory>../../</directory>
+			<exclude>
+				<directory suffix=".php">../../tests/</directory>
+				<directory suffix=".php">../../tmp/</directory>
+				<directory suffix=".php">../../vendor/</directory>
+				<directory suffix=".php">../../bower_components/</directory>
+				<directory suffix=".php">../../resources/</directory>
+				<directory suffix=".php">../../initialisation/</directory>
+				<directory suffix=".php">../../src/assets/resources/</directory>
+			</exclude>
+		</whitelist>
+	</filter>
+
+</phpunit>
