+++ tests/phpunit/unit-tests/test-notices.php
